WIP very basic PoC of high performance profile
This is based on features we already have or will have
in very near future. Some more interesting features
will require changes to how VMI presets are evaluated.

Please do not merge yet, I am currently using this to
collect all the config pieces we think are needed for
high performance VMs.

This currently depends on couple of PRs:

- https://github.com/kubevirt/kubevirt/pull/1421
- https://github.com/kubevirt/kubevirt/pull/1381
- https://github.com/kubevirt/kubevirt/pull/1385

Signed-off-by: Martin Sivak <msivak@redhat.com>