Implement osinfo lookup
This adds osinfo lookup source that makes it possible
to use libosinfo information in templates.

There are two ways to reference an attribute:
- object style - lookup().name
- dict style - lookup()["name"]

Arrays can be accessed using numerical indices like this:
lookup('osinfo', 'fedora15')["minimum_resources.0.ram"]

Beware though: Ansible has a bug [1] that makes it
impossible to use this lookup source from Ansible files
like this: "{{ lookup(..) }}". Ansible / Jinja2 always
converts the result to AnsibleUnsafeText intead of keeping
it in the original native type and that breaks the proxy
used to access all the result attributes.

[1] https://github.com/ansible/ansible/issues/34595#issuecomment-356091161