windows: templates: enable synic flags
Kubevirt gained support for pretty much all the hyper-v related flags
exposed in libvirt 5.1.
In this patch we update the common templates to use the synic flags:

- vpindex: needed to enable SyNIC, and also for few optimization
     on overcommitted environments (ipi, tlbflush), which are planned
     to be added soon.
- synic: needed to enable synictimer
- synictimer: reportedly improves performance significantly
     on windows >= 10

Please note that we DO NOT assume or require the HyperV strict
checking[1].
We leave that to the cluster admin or to some other opinionated setups.

+++

[1]: https://kubevirt.io/user-guide/docs/latest/creating-virtual-machines/guest-operating-system-information.html#hyperv-optimizations

Signed-off-by: Francesco Romani <fromani@redhat.com>