Allow searching in osinfo db when retrieving data
This change enhances the data retrieval language to allow
selecting the right record from multiple architecture
dependent data sets.

For example: The minimal memory requirements are
specified per architecture. The template can access
the right value like this:

{{ lookup('osinfo', osinfoname)["minimum_resources.architecture=x86_64|all.ram"] }}

Pay attention to the architecture part. Whenever an equals sign
is encountered in the field name a search is performed.
It is expected the current level is a list and the key=value pair
is used to find the first elements that has key equal to value.

There are two enhancements to the single search.
- A value can contain the | (pipe) character and in such case
  the search treats the string as a list of alternative values.
  The pipe has the OR meaning and the lowest priority
- There can be many key=value pairs separated by comma
  key1=value1,key2=value2 and this is treated as logical AND.

Signed-off-by: Martin Sivak <msivak@redhat.com>