Add fixed version suffix -v0.7.0 to all templates
This is a slight update to the recent patch that removed release
specific suffixes from template names. That fix forgot to take
into account the existing deployments that were done using SSP
operator and HCO.

The last version released via those channels was -v0.7.0 and
so to make upgrades from that release possible, I am readding
the suffix statically.

This can be interpreted as a version in which the template was
first officially released.

v0.7.0 is also the first version that appeared in a product
that fully supports upgrades. All versions prior that were
always static and had no upgrade guarantees.

Signed-off-by: Martin Sivak <msivak@redhat.com>