Add check for min memory consistency in generated templates
Currently the minimum memory requirements set in the "validation" part of the generated template
might actually be smaller than the requirements of newest minor version of the os,
this script will through an error for such a template.

IMHO a better solution should be implemented in the future, but the obvious one - setting the
minimal memory requirement to the actual OS requirement will not do since we need to keep naming consistency,
i.e. "tiny" template means a specific memory requirement throughout the system.

Signed-off-by: borod108 <bodnopoz@redhat.com>