osinfo: only try to load "overrides" if the directory exists
aa7b15342ba1f533b4f3ddb3c0480f1c1503667f started loading the content
from "osinfo-db-override", but without the precaution of checking
whether this directory exists or not.

As git removes a directory whenever it's empty, we can't safely assume
the directory will be there, which causes a breakage.  In order to avoid
that, let's just check whether the directory exists before trying to
load its content, and we should be good to go after that.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>