Validate that this repository does not contain offensive language (#365)
Add a check to the sannity test that fails if there is a usage
of offensive language in one of the files.

The script forbid the use of the words `master`, `slave`,
`blacklist` and `whitelist`

Since `master` is very common word as a name of default
branches or as a node name in clusters, the script allows
the usage of it if this word is part of a URL or a path.
The script does not check the directory
.github/workflows
that may contains a lot if usage of the word `master`, which
refers to the master branch.

Signed-off-by: Dominik Holler <dholler@redhat.com>