Use containerdisk urls provided by osinfo-db
With this change containerdisk urls for the templates are loaded from
osinfo-db instead of specifying them manually (where applicable).

Overrides for centos-stream[8|9] are still needed because of CentOS
still having no stable image urls. Containerdisks are added there for
now too. The image urls are updated to the latest version.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>