Use scsi as the default bus in VM templates
Virtio-blk disks consume addresses on the PCI bus and this is an important limitation. Since performance parity has been shown between virtio-scsi and virtio-blk it makes sense to standardize the former.

This commit updates the VM templates so, by default and when possible, disks are created using the SCSI bus.

Signed-off-by: Alvaro Romero <alromero@redhat.com>