fix: make unit-tests can be called more than once
In order to run the unit tests multiple times, the kubevirt
namespace will be removed before new templates are created. This
method ensures that unit tests can be executed successfully, even
if the templates already exist in the cluster.

If the namespace does not exist, the 'oc delete' command would
fail, hence the use of '|| true' to handle this scenario.

Signed-off-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>