fix: patterns in expect script and adjust timeout/exit conditions
The issue identified was that the string "1" is recognized as a
pattern. The latest version of CentOS Stream 9 includes this
"pattern" in its boot sequence.

This commit also reveals some errors that failed silently,
specifically related to highperformance configurations. For now,
the plan is to disable these configurations and address the issues
in a future pull request.

Signed-off-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>